pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=22.0",
    "requests>=2.32.3",
    "yt-dlp>=2024.3.10",
]
